db fix & file move
