+++ packages/backend/src/fetchdatabase.ts
@@ -47,8 +47,8 @@ export const fetchDatabase = async (usingApproved: boolean = false) => {
     '0loyalty',
     '0oracle_text',
     '0flavor_text',
-    'artists',
     '0image',
+    'artists',
     'tags',
     '1mana_cost',
     '1supertypes',
