+++ packages/frontend/src/hellfall/stringToMana.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { HCCardSymbol, HCColors } from '@hellfall/shared/types';
-import { sameColors } from './opComps';
+import { sameColors } from './filters/opComps.ts';
 import { getPipsData } from '@hellfall/shared/services/pipsService.ts';
 import { withBasePath } from '../basePath.ts';
